% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-anim-pattern-rotation.R
\name{create_anim_pattern_rotation}
\alias{create_anim_pattern_rotation}
\title{Create a simple rotation animation for an SVG pattern}
\usage{
create_anim_pattern_rotation(
  duration = 16,
  cx = 0,
  cy = 0,
  repeatCount = "indefinite",
  ...
)
}
\arguments{
\item{duration}{time for full rotate (seconds). default: 16}

\item{cx, cy}{centre of rotation. default (0, 0)}

\item{repeatCount}{default: 'indefinite'}

\item{...}{other arguments passed to \code{minisvg::stag$animateTransform()}}
}
\value{
minisvg SVGElement
}
\description{
This animation targets the \code{patternTransform} attribute, and thus onlyl
applies to patterns.
}
\details{
If you want to animate the rotation of an arbitrary SVG object, see
\link{create_anim_rotation}.
}
\examples{
\dontrun{
# Create the animation
anim1 <- create_anim_pattern_rotation(id = 'hello')

# Create a pattern and append the animation to the pattern
pat   <- svgpatternsimple::create_pattern_stripe(id = 'stripe')
pat$append(anim1)

# Create an SVG document with a rectangle, add the pattern to the
# document definitions, and use the pattern in the rectangle
doc <- minisvg::svg_doc()
doc$defs(pat)
doc$rect(x=0, y=0, width='100\%', height='100\%', fill=pat)

# view the code and render the SVG
doc
doc$show()
}

}
