% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-anim-pulse-line.R
\name{create_anim_pulse_line}
\alias{create_anim_pulse_line}
\title{Create a simple pulsing animation in which the width of the line varies}
\usage{
create_anim_pulse_line(
  size1 = 0,
  size2 = 10,
  duration = 4,
  repeatCount = "indefinite",
  ...
)
}
\arguments{
\item{size1, size2}{the minimum and maximum size of the strokes}

\item{duration}{time for full cycle (seconds). default: 4}

\item{repeatCount}{default: 'indefinite'}

\item{...}{other arguments passed to \code{minisvg::stag$animate()}}
}
\value{
minisvg SVGElement
}
\description{
Create a simple pulsing animation in which the width of the line varies
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the animation
pulse <- create_anim_pulse_line(size2 = 50)

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  fill   = "lightblue",
  stroke = 'black',
  pulse
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc
doc$show()
}
}
