% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-anim-rotation.R
\name{create_anim_rotation}
\alias{create_anim_rotation}
\title{Create a simple rotation animation}
\usage{
create_anim_rotation(
  duration = 16,
  cx = 0,
  cy = 0,
  repeatCount = "indefinite",
  ...
)
}
\arguments{
\item{duration}{time for full rotate (seconds). default: 16}

\item{cx, cy}{centre of rotation. default (0, 0)}

\item{repeatCount}{default: 'indefinite'}

\item{...}{other arguments passed to \code{minisvg::stag$animateTransform()}}
}
\value{
minisvg SVGElement
}
\description{
Create a simple rotation animation
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the animation
rot   <- create_anim_rotation(cx = 200, cy = 200)

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  fill   = "lightblue",
  stroke = 'black',
  rot
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc$show()
}

}
