% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-anim-stroke-render.R
\name{create_anim_stroke_render}
\alias{create_anim_stroke_render}
\title{Create an animation for the rendering of a stroke around an object.}
\usage{
create_anim_stroke_render(
  offset = 2000,
  duration = 4,
  repeatCount = "indefinite",
  ...
)
}
\arguments{
\item{offset}{The length of the offset motion. this will need to be adjusted
depending on the overall stroke length. default: 2000}

\item{duration}{time for full cycle (seconds). default: 4}

\item{repeatCount}{default: 'indefinite'}

\item{...}{other arguments passed to \code{minisvg::stag$animate()}}
}
\value{
minisvg SVGElement
}
\description{
This animation will only show if the object it applies to (a) has a visible
stroke, and (b) has the \code{stroke-dasharray} attribute set.
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the animation
stroke_render <- create_anim_stroke_render()

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  fill   = "lightblue",
  stroke = 'black',
  stroke_width = 5,
  stroke_dasharray = 2000,
  stroke_render
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc
doc$show()
}
}
