% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svganim.R
\docType{package}
\name{svganim}
\alias{svganim}
\title{svganim: A package of parameterised SVG animations built using \code{minisvg}}
\description{
svganim: A package of parameterised SVG animations built using \code{minisvg}
}
